+++ src/idunno.Bluesky/Drafts/DraftPost.cs
@@ -304,7 +304,7 @@ namespace idunno.Bluesky.Drafts
             {
                 if (Labels is not null)
                 {
-                    return Labels.Contains(SelfLabelNames.Porn);
+                    return Labels.Contains(SelfLabelValues.Porn);
                 }
                 else
                 {
@@ -317,11 +317,11 @@ namespace idunno.Bluesky.Drafts
                 Labels ??= new SelfLabels();
                 if (value)
                 {
-                    Labels.AddLabel(SelfLabelNames.Porn);
+                    Labels.AddLabel(SelfLabelValues.Porn);
                 }
                 else
                 {
-                    Labels.RemoveLabel(SelfLabelNames.Porn);
+                    Labels.RemoveLabel(SelfLabelValues.Porn);
                 }
             }
         }
@@ -337,7 +337,7 @@ namespace idunno.Bluesky.Drafts
             {
                 if (Labels is not null)
                 {
-                    return Labels.Contains(SelfLabelNames.Sexual);
+                    return Labels.Contains(SelfLabelValues.Sexual);
                 }
                 else
                 {
@@ -350,11 +350,11 @@ namespace idunno.Bluesky.Drafts
                 Labels ??= new SelfLabels();
                 if (value)
                 {
-                    Labels.AddLabel(SelfLabelNames.Sexual);
+                    Labels.AddLabel(SelfLabelValues.Sexual);
                 }
                 else
                 {
-                    Labels.RemoveLabel(SelfLabelNames.Sexual);
+                    Labels.RemoveLabel(SelfLabelValues.Sexual);
                 }
             }
         }
@@ -373,7 +373,7 @@ namespace idunno.Bluesky.Drafts
                     return false;
                 }
 
-                return Labels.Contains(SelfLabelNames.GraphicMedia);
+                return Labels.Contains(SelfLabelValues.GraphicMedia);
             }
 
             set
@@ -381,11 +381,11 @@ namespace idunno.Bluesky.Drafts
                 Labels ??= new SelfLabels();
                 if (value)
                 {
-                    Labels.AddLabel(SelfLabelNames.GraphicMedia);
+                    Labels.AddLabel(SelfLabelValues.GraphicMedia);
                 }
                 else
                 {
-                    Labels.RemoveLabel(SelfLabelNames.GraphicMedia);
+                    Labels.RemoveLabel(SelfLabelValues.GraphicMedia);
                 }
             }
         }
@@ -404,7 +404,7 @@ namespace idunno.Bluesky.Drafts
                     return false;
                 }
 
-                return Labels.Contains(SelfLabelNames.Nudity);
+                return Labels.Contains(SelfLabelValues.Nudity);
             }
 
             set
@@ -413,11 +413,11 @@ namespace idunno.Bluesky.Drafts
 
                 if (value)
                 {
-                    Labels.AddLabel(SelfLabelNames.Nudity);
+                    Labels.AddLabel(SelfLabelValues.Nudity);
                 }
                 else
                 {
-                    Labels.RemoveLabel(SelfLabelNames.Nudity);
+                    Labels.RemoveLabel(SelfLabelValues.Nudity);
                 }
             }
         }